fix(#146): anomaly_scores / changepoints / drift_reports скоупированы по project_id
## Проблема

Три ML-таблицы не содержали project_id — история аномалий, change-point'ов
и drift-отчётов была глобальной. Дашборд пользователя A показывал ML-данные
пользователя B. Кроме того, старые PRIMARY KEY (ts, table_name) и
(table_name, column_name) перетирали строки разных проектов при upsert.

## Схема и миграция

### scripts/metrics_schema.sql, scripts/timescale_schema.sql
- `anomaly_scores`: добавлен `project_id TEXT NOT NULL DEFAULT 'legacy'`,
  PK изменён с `(ts, table_name)` на `(project_id, ts, table_name)`.
- `changepoints`: PK изменён с `(ts, table_name, metric_name)` на
  `(project_id, ts, table_name, metric_name)`.
- `drift_reports`: PK изменён с `(table_name, column_name)` на
  `(project_id, table_name, column_name)`.
- Добавлены индексы `(project_id, ts DESC)` / `(project_id, computed_at DESC)`
  для эффективной фильтрации по проекту.

### app/metrics_storage.py — _migrate_existing_schema()
Добавлена `_migrate_project_scoped_ml_tables(engine)` — автоматически
мигрирует существующие БД при старте приложения:

- **SQLite**: CREATE TABLE {table}__new → INSERT SELECT с `'legacy'` backfill
  → DROP TABLE {table} → RENAME. Всё в одной транзакции. Идемпотентно:
  проверяет наличие project_id в колонках и соответствие нового PK перед
  запуском. DROP TABLE IF EXISTS {table}__new защищает от прерванной
  предыдущей миграции.
- **Postgres**: ALTER TABLE ADD COLUMN project_id DEFAULT 'legacy' (если
  отсутствует) → DROP CONSTRAINT → ADD PRIMARY KEY с новыми колонками.
- Код миграции описан через `_PROJECT_SCOPED_ML_TABLES` dict — добавить
  четвёртую таблицу в будущем стоит одна запись.

## Storage API

Все функции чтения и записи получили `project_id: str = "legacy"`:

- `save_anomaly_scores(rows, project_id)` — upsert-ключ теперь
  `(project_id, ts, table_name)`.
- `get_anomaly_scores(table_name, project_id, window)` — WHERE project_id.
- `save_changepoints(rows, project_id)` — conflict_columns, delete_stmt
  и cluster_query (dedupe ±72 ч) фильтруются по project_id: A не может
  вытеснить change-point B.
- `get_changepoints(table_name, ..., project_id)` — WHERE project_id.
- `save_drift_reports(table_name, rows, project_id)` — DELETE + INSERT
  scoped по project_id.
- `get_drift_report(table_name, project_id)` — WHERE project_id.
- `_fetch_anomalies_by_ts(project_id, window)` — history page больше
  не считает чужие аномалии. Убран комментарий «NB: anomaly_scores
  still global».

## ML-модули

### ml/anomaly_detector.py
- `_load_features`, `train`, `_load_model`, `score_table`,
  `feature_breakdown`, `retrain_all` — все получили `project_id="legacy"`.
- `_model_path(table, project_id)`: для `project_id != "legacy"` файл
  сохраняется как `models/{project_id}__{table}__anomaly.joblib`. Legacy
  path неизменён для обратной совместимости.
- `retrain_all` и `detect_all` / `compute_and_store_drift_all` принимают
  `tables: Iterable[str] | None` — если передан, обходит именно этот
  список вместо `list_tables()` из target DB.

### ml/changepoint.py
- `detect_changepoints(table, metric, window_days, project_id)` —
  читает метрики из нужного проекта.
- `detect_all(metrics, window_days, project_id, tables)` — передаёт
  project_id в detect_changepoints и save_changepoints.

### ml/drift.py
- `_load_distributions(table_name, since, project_id)` — убран хардкод
  `"legacy"`, был уже с `WHERE project_id = :project_id` (#53).
- `compute_drift(table_name, baseline_days, project_id)`.
- `compute_and_store_drift_all(project_id, tables)`.

## Dashboard / API / LLM

### app/dashboard.py
- `_ml_last_runs(project_id)` — MAX(ts)/MAX(detected_at)/MAX(computed_at)
  фильтруются по project_id. Вызов: `_ml_last_runs(_current_project_id())`.
- `schema_view`, `table_detail` — `get_drift_report(name, _current_project_id())`.

### app/api.py
- `/api/drift/<table>` — `get_drift_report(table, _current_project_id())`.
- `/api/changepoints/<table>` — `get_changepoints(..., project_id=...)`.
- `/api/anomalies/<table>` — `get_anomaly_scores(..., project_id=...)` +
  `feature_breakdown(..., project_id=...)`.
- `/api/explain` — `explain_anomaly(..., project_id=_current_project_id())`.

### app/llm.py
- `_build_prompt` и `explain_anomaly` — все reads (get_metrics, get_changepoints,
  get_anomaly_scores) передают project_id. Убран комментарий «#53: global».

## warmup_ml + Makefile

### scripts/warmup_ml.py
- `--project-id` CLI аргумент.
- `_tables_from_metrics(project_id)` — `SELECT DISTINCT table_name FROM
  metrics WHERE project_id = :pid`. Не требует live-connection к target DB.
- Для `project_id="legacy"` при пустых метриках `table_scope=None` →
  fallback на `list_tables()` (обратная совместимость). Для non-legacy
  при пустых метриках `table_scope=[]` → ничего не тренируется.

### Makefile
- `warmup-ml` передаёт `--project-id $(PROJECT_ID)`.

## Seeder

### scripts/seed_metrics_db.py
- `_purge_existing`: `_SCOPED` теперь включает `anomaly_scores`,
  `changepoints`, `drift_reports`. Scoped DELETE при non-legacy reset.
- Убрана ссылка на `(#146)` из WARNING-лога.
- Попутно: `int(round(...))` → `round(...)` (RUF046, pre-existing).

## Тесты

- `test_changepoints_are_scoped_by_project` — одинаковый (table, ts)
  в двух проектах хранится независимо.
- `test_changepoint_dedupe_does_not_cross_project` — overlapping change-points
  в разных проектах не вытесняют друг друга.
- `test_drift_reports_are_scoped_by_project`.
- `test_anomaly_scores_are_scoped_by_project`.
- `test_history_anomalies_are_scoped_by_project` — proj-a не видит
  аномалии proj-b в build_history_aggregate.
- `test_migrates_legacy_ml_tables_to_project_scoped_pk` — создаёт
  SQLite с дореформенной схемой, перезапускает engine, проверяет PK
  через _sqlite_pk_columns и читает backfill-строки.

+++ tests/test_seed_metrics_db.py
@@ -186,7 +186,7 @@ def test_drift_factor_monotonic_after_onset():
 def test_categorical_buckets_baseline_when_no_drift():
     out = _categorical_buckets(progress=1.0, drift_amount=0.0)
     weights = [b["count"] for b in out]
-    expected_total = sum(int(round(w / sum(CATEGORICAL_BASELINE_WEIGHTS) * 1000))
+    expected_total = sum(round(w / sum(CATEGORICAL_BASELINE_WEIGHTS) * 1000)
                          for w in CATEGORICAL_BASELINE_WEIGHTS)
     assert sum(weights) == expected_total
     # Бакеты соответствуют именам.
